Analytics: Duplicate to a new folder
We will be reverting 215ff5f9a50ae495e4c1aefda470151985cd8063 and 83c7cfa50eb4aba35a657a2fa525e3a3520b78bd

The reason we are reverting is that we want two separate layouts
for the AppRouter and PageRouter of Next.

This we ensure no interaction between the two. However, this does duplicate
the code. We will moving all pages to the new AppRouter in the future.