Layout: Duplicate
We want to duplicate the frontend/layouts/default/index.tsx file.

The reason is that we plan to revert the layout changes here and we want to keep the original layout file.

We are working to ensure the App Router / Pages Router have their own layout files and
any changes to support the newer APIs of next.

The goal is to avoid any interaction at the cost of duplication.