feat(sentry): cache main branch data
Caches the most recent HEAD commit sha of the main branch. This way, we don't need to talk to GitHub _every_ time we report an error.

Having slightly stale commit data is ok here. The intent of the current_production tag is to filter out old deployments when searching errors in Sentry. 99% of old deployments will be tagged correctly.