iFixit API: Support Client only functions
Next (App Router) requires all functions which would be used on the
client to be in the files marked with 'use client'.

I moved the two functions which are not able to run as server components
to the client file.

You can import / re-export them from a server file with no issue.

Thus, no external changes occurred.