Mobile TOC: Only show once we have scrolled into content
Now that we may have an active element in view on load, we don't want to
have the mobile TOC be on top of the header.

We can trigger the toc to show after we have scrolled into the container.

For clarity, the previous commit changed the behavior of which element was active.
On load we would never have an active element in the prior code, but now we do.