Mobile TOC: Handle display state better
On Desktop, the mobile TOC was rendered, but hidden.
This caused the dektop view to run the scroll handler for the mobile TOC
If the scroll handler did a set state, the scroll would be stopped
even on desktop view!

On mobile, the TOC is only visible on these setstate calls, but it's always visible on dektop.

The fix here was to stop running the scroll handler if the mobile TOC is display none.