Reorganize functions for handling shopify variant uri
The shopify variant id is a piece of data that is stored in various ways. Its either by itself, in a shopify product variant uri? or in that uri and encoded. This adds functions with better names to more clearly convert between the three stages. It then also uses those functions where appropriate.