Minor: Simplify useEffect hook
We only want to scroll to the record element once, when the component is mounted.
Any subsequent changes to the record should not trigger a scroll.

Therefore, we use an empty dependency array to ensure that the effect only runs once,
and we disable the react-hooks/exhaustive-deps ESLint rule to avoid a
warning about the missing record dependency.