WorkBench: Get data over API.
I think we will need to update the API to return more data.

The given figma has more info about a wiki.

It's also possible we get the info from aloglia,
but that means another hop to get all the data needed to render the page

One note: This code is still relying on clientside loading
Thus there is still CLS.

I care more about getting the data and testing at this point that solving the
CLS.