Client: Split out json functions
95% of our API we expect an exception if not 200 or a 200 with json.

We should allow devs to access the response, but also allow a nice default handler

Previously, there was an optional callback on how to process the response,
but now I have split this into two funcs.

One which calls the default respinse handler for json and one which hands
the full response object to the dev.

One note: I left one callsite broken on this commit and it will be fixed in the next commit.