Global Fetch Reporting: Bypass on our own API
We have cases where our own API would return a 422 and we don't want to hear about it in Sentry
We also have cases where we want an external API to ping Snetry on the same status code.

After much discussion we landed on always ignoring the global fetch error handling on our API
in favor of custom logic.

i.e. We will still ping Sentry, but only if a dev opts into the getJson function