Minor: simplify code
I'd view this with no whitespace

The code behaves the same except for one thing:

We now remove the header which we pass around as signal to bypass the error checking.

It's unneeded on the request and saves bytes, but also I can see it being set then not in the request magicaly to devs.

FOr now, Let's try this approach