Fetch Error: Set more Sentry details
This adds more details to the Sentry error when a fetch() call fails.

Since the json() method and the text() method on the response object
are async, we can't call them in the FetchError constructor.

In fact, we need to ensure the sentry details are set before triggering a
throw.

This commit also adds a new function, throwErrorWithSentryDetails, which
does just that.

We add a new sentry scope, await the json() or text() method, and then
set the sentry details. We immediately throw the error after that.

This ensures the scope is not mutated after the error is thrown, but caught
further up the stack.

This commit also adds a few more tags to the Sentry error, including the
request URL, the response status code, and the response status text.

For some extra data, we do our best to include the request and response
objects in the Sentry contexts