Sentry: Allow setting tags and contexts on an error
It ends up Sentry does not work as expected:
https://github.com/getsentry/sentry-javascript/issues/2948

In effect this works just fine:

Sentry.withScope((scope) => {
   scope.setTags({ foo: 'bar' });
   scope.setContext('foo', { bar: 'baz' });

   Sentry.captureException(new Error('foo'));

});

While this will not add any of the scope data to the error:

Sentry.withScope((scope) => {
   scope.setTags({ foo: 'bar' });
   scope.setContext('foo', { bar: 'baz' });

   throw new Error('foo');
});

We often really want to throw an error while also having extra data for sentry.

This commit solves that by having a new class which has data for sentry
and adds a global event processor which will add the data to the error.

We implemented this differently in the monorepo, but I think this is simpler.
Ref: https://github.com/iFixit/ifixit/pull/37791/commits/f26d595fd10308f0563344dca4c8db441894942c