withCache: bubble up errors
I have some concerns about this change.

It's unclear why we chose to shim the error into a result object.

The code was added here:
https://github.com/ifixit/react-commerce/commit/da3419bf7c6bebb62ca55953226de9efc70af743#diff-464ca1bbc3a97537d4a05baff86638ced1b605dc18f56bbc4b3b0d8a3140c19aR136

Before this commit, an existing error from a fetch would simply stop and be hidden away.

We ran into this when

await storefront.findProduct({
            handle: 'test-handle',
         });

failed due to a Shopify outage.

We actually threw errors with debugging info. However,
these errors never made it to sentry since we catch them.

There is another way to handle these errors:

- explictly capture the exception and keep the result object as the return

Opinions are welcome.