fix(product list): omit default layout title
Fixes a bug where the title would flash correctly as the page loaded, then change to "iFixit" suddenly (which is set in DefaultLayout). This is probably a problem with Algolia, because as soon as you interact with any search facet or the search box, the page title changes back to the product list meta title.

This is a bandaid solution, it would be better to fix the root of the problem in our Algolia configuration.