Next: Use SWC
This small change took some digging.

1) We decided to try setting this to false in hopes of fixing two issues

https://github.com/iFixit/react-commerce/issues/1797
https://github.com/iFixit/react-commerce/pull/1846

The first issue was fixed (1797) while the second issue was not (1846).

2) Next 14 fails to build with terser enabled.

This is because terser doesn't support node some out the new tsx used by Next 14.
The error is:

../pages/_document.js from Terser
Unexpected token: punc ({) [../pages/_document.js:675,10]

My best guess from some googling is this is due to the class static keyword:
https://github.com/terser/terser/issues/1160

However, this is fixed simply by using the SWC minifier (default) instead of terser.

3) Now with Next 14 the docs claim we shouldn't need terser anymore: https://nextjs.org/docs/architecture/nextjs-compiler
In fact, Terser will not be supported in Next 15 slated for release next year.