Test: Ensure we end up on the correct page
For some reason a slight behavior has changed such that the url we load after hitting
the logout button is `/` instead of `/tools`.

The PHP code looks for a redirect param, then the HTTP_REFERER then falls back to `/Guide`.
https://github.com/iFixit/ifixit/blob/178a3d4a7070f989898ea4196e06d3f6eb99583f/Guide/UI/User/logout.php#L20-L22

AFAICT this would mean the HTTP_REFERER is now `/` instead of `/tools`.

I'm not sure why this is happening, but I'm going to update the test to reflect the
current behavior.

Overall we are asserting that logout worked here.