Product Images: incorporate fallbacks
Products may now have fallback images (only one right now, but using an
array made integration easier). These are only to be used when a variant
has no other image to represent it. They are identified by (altText ===
'*fallback')

Here we extract them into a new property and make use of them wherever
we are using product.images.

We can't inject copies of these fallback images into product.images
otherwise the gallery would show one copy of the fallback for each
variant that uses it.

fallback images won't always be present and we've dealt with that here.
Our schema validation on the cache should protect us against this
change.

Note: We are filtering out fallback images from the normal
product.images because in that array, variantId === null implies it's
valid for all variants and that's not true for fallback images.