Troubleshooting: Show Related Problems appropriately
When we have the `extended-related-problems` flag enabled, we have extra
problems potentially included in the `Related Problems` list
(confusingly, these come from the `relatedProblems` key and are added to
the existing `linkedProblems` list). This ensures that we update the
`hasRelatedPages` boolean to accurately reflect the state of the
`Related Problems` section.