fix(config): drop legacy product list redirects
Drops redirects for legacy device product list slugs. We're not confident that these urls are still seeing traffic, and these redirects are beginning to conflict with new item types which is a problem. There's some debate about whether we should continue supporting the legacy tool categories, so they were excluded from this change.