gh actions: apply masking option to ecr login
There is a warning about the 'mask-password' option for the
"amazon-ecr-login" action. If ran in debug mode, the password will be
outputted. By default this option is not enabled as the expectation is
to run the service after the deployment but as a secondary job. As
such the password cannot be masked as it is needed to run the service.

https://github.com/aws-actions/amazon-ecr-login#run-an-image-as-a-service

However, this is not the workflow we have set up, therefore we need to
enable the option to mask the password.

Reference: https://github.com/iFixit/vigilo/pull/19