refactor: Create new footer for app router
While moving the product page to the app router one of the challenges is to customize the region menu in the footer using product links. Due to the nature of the app router this is not doable via React context, since the layout does not change on page navigation and product data is contained within it.
In this scenario the solution that I have found is to use parallel routes, but this requires some change in the footer component, thus we're creating a new footer component that is specific to the app router